change chunk to []byte, start impl table contructor

+++ internal/table/chunk_test.go
@@ -7,7 +7,7 @@ import (
 
 func TestChunkQuery(t *testing.T) {
 	chunk := chunk{
-		data: [][]int{
+		data: [][]byte{
 			{
 				1, 2, 3, 4, 5,
 			},
